backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
mediapipe==0.10.14
opencv-python-headless==4.9.0.80
numpy==1.26.4
websockets==12.0
python-multipart==0.0.9
sqlalchemy==2.0.29
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
